README.md updates & Fix build (split requirements)
Update README.md with more relevant and informative data. so as split
the requirements to -dev and -prod for development / testing
requirements and production requirements.
The thing is, the production environment requires scipy which takes A
LOT of time to install, this causes Travis-CI executor to hang up for
more than 20mins and hence fail it.
For the meantime, we dont need these libaries on Travis-CI tests.